Narona_ASI/requirements.txt
google-generativeai>=0.8.0
google-genai>=0.8.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.14
gpiozero>=2.0
RPi.GPIO>=0.7.1
smbus2>=0.4.3
requests>=2.31.0
Pillow>=10.0.0
